Turn ChannelTracker into a HOC
